scripts/requirements.txt
# Python Dependencies untuk Sentiment Analysis TA
# Install dengan: pip install -r requirements.txt

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Natural Language Processing
nltk>=3.6.0
Sastrawi>=1.0.1

# Machine Learning
scikit-learn>=0.24.0
imbalanced-learn>=0.8.0

# Database
mysql-connector-python>=8.0.0

# Optional: Untuk Jupyter Notebook
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
